// src/pages/themes/data.d.ts
export type Theme = {
  _id?: string
  title?: string
  url?: string
  image?: string
  data?: any
}

export type Tab = {
  _id?: string
  title?: string
  value?: string
}

export type ListThemes = {
  status: boolean
  data?: Theme[]
}

export type ListTabs = {
  status: boolean
  data?: Tab[]
}

export type ThemeT = {
  listTheme: ListThemes
  listTab: ListTabs
}